Moved `sail configure` command to `sail set pat`
